Rename variable into references
