+++ packages/cypher-builder/src/clauses/Create.ts
@@ -18,8 +18,8 @@
  */
 
 import type { CypherEnvironment } from "../Environment";
-import type { NodeRef } from "../variables/NodeRef";
-import type { Param } from "../variables/Param";
+import type { NodeRef } from "../references/NodeRef";
+import type { Param } from "../references/Param";
 import { Pattern } from "../Pattern";
 import { SetClause } from "./sub-clauses/Set";
 import { Clause } from "./Clause";
