+++ packages/cypher-builder/src/clauses/With.ts
@@ -21,8 +21,8 @@ import type { CypherEnvironment } from "../Environment";
 import { Projection } from "./sub-clauses/Projection";
 import type { Expr } from "../types";
 import { compileCypherIfExists } from "../utils/compile-cypher-if-exists";
-import type { Literal } from "../variables/Literal";
-import type { Variable } from "../variables/Variable";
+import type { Literal } from "../references/Literal";
+import type { Variable } from "../references/Variable";
 import { Clause } from "./Clause";
 import { WithOrder } from "./mixins/WithOrder";
 import { WithReturn } from "./mixins/WithReturn";
