+++ packages/cypher-builder/src/clauses/sub-clauses/Remove.ts
@@ -19,7 +19,7 @@
 
 import { CypherASTNode } from "../../CypherASTNode";
 import type { CypherEnvironment } from "../../Environment";
-import type { PropertyRef } from "../../variables/PropertyRef";
+import type { PropertyRef } from "../../references/PropertyRef";
 
 // TODO: Remove label
 export type RemoveInput = Array<PropertyRef>;
