+++ packages/cypher-builder/src/references/Literal.ts
